Make the project compile
Well LLVM is utter bullshit when it comes to stability.
I had to struggle between mailinig lists and different version of the
LLVM to get this working so ... at the end it complies
Also I made some drafting to how the datastructures I will be using will
look like. I won't be surprised if that thing didn't work at the end.